+++ device_inductance/__init__.py
@@ -11,11 +11,34 @@ from device_inductance import model_reduction, contour, sensors
 from device_inductance.device import DeviceInductance, TypicalOutputs
 from device_inductance.coils import Coil, CoilFilament
 from device_inductance.structures import PassiveStructureFilament
+from device_inductance.logging import log, logger_is_set_up, logger_setup_default
+
 
 def load_default_ods() -> ODS:
-    """Load an example ODS file in the format required by device_inductance."""
+    """
+    Load an example ODS file in the format required by device_inductance.
+
+    The example differs from real SPARC configurations in at least the following ways:
+      * Coil number of turns and resistances are obfuscated
+      * Actual magnetics sensors are replaced with mockup examples
+      * The example description may be arbitrarily out-of-date, as it is not updated regularly
+    """
+
+    if not logger_is_set_up():
+        logger_setup_default()
+
+    log().warning("""
+Loading example device description.
+The example differs from real SPARC configurations in at least the following ways:
+    * Coil number of turns and resistances are obfuscated
+    * Actual magnetics sensors are replaced with mockup examples
+    * The example description may be arbitrarily out-of-date, as it is not updated regularly
+                  """)
+
     # NOTE: This should be rewritten to use importlib once omas supports loading raw text
-    ods_filename = Path(__file__).parent / "../examples/OS_SPARC_Device_Description.json"
+    ods_filename = (
+        Path(__file__).parent / "../examples/OS_SPARC_Device_Description.json"
+    )
     with open(ods_filename) as f:
         ods = load_omas_json(f)
 
@@ -47,6 +70,7 @@ def load_default_ods() -> ODS:
 
     return ods
 
+
 def typical(
     ods: ODS,
     extent: tuple[float, float, float, float] = (0.0, 0.0, 0.0, 0.0),
@@ -54,6 +78,7 @@ def typical(
     max_nmodes: int = 40,
     model_reduction_method: Literal["eigenmode", "stabilized eigenmode"] = "eigenmode",
     show_prog: bool = True,
+    plasma_coil_force_method: Literal["tables", "mask"] = "mask",
 ) -> TypicalOutputs:
     """
     Generate a typical set of outputs,
@@ -70,6 +95,10 @@ def typical(
         dxgrid: [m] Spatial resolution of computational grid
         max_nmodes: Maximum number of structure modes to keep. Defaults to 40.
         show_prog: Whether to show terminal progress bars. Defaults to True.
+        plasma_coil_force_method: Whether to interpolate B-field on the fully-realized mesh tables,
+                                  or do direct filament calculations from points inside the limiter mask.
+                                  Defaults to "mask", which is faster and uses less memory, but only includes
+                                  nonzero entries inside the limiter, which requires a valid limiter geometry.
 
     Returns:
         A fully-computed set of matrices and tables covering the needs of a typical workflow
@@ -81,6 +110,7 @@ def typical(
         dxgrid=dxgrid,
         model_reduction_method=model_reduction_method,
         show_prog=show_prog,
+        plasma_coil_force_method=plasma_coil_force_method,
     )
 
     out = TypicalOutputs(